# web/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MemEvolve API",
    "description": "Memory-enhanced proxy and evolution framework for OpenAI-compatible LLM APIs",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:11436",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/memory/search": {
      "post": {
        "tags": ["Memory"],
        "summary": "Search Memory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemorySearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemorySearchResult"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/evolution/record-request": {
      "post": {
        "tags": ["Evolution"],
        "summary": "Record API Request Metrics",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "latency_ms": { "type": "number" },
                  "tokens": { "type": "integer" },
                  "success": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Recorded" }
        }
      }
    },
    "/evolution/record-retrieval": {
      "post": {
        "tags": ["Evolution"],
        "summary": "Record Memory Retrieval Metrics",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "retrieval_time_ms": { "type": "number" },
                  "results_count": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Recorded" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MemorySearchRequest": {
        "type": "object",
        "required": ["query"],
        "properties": {
          "query": { "type": "string" },
          "limit": { "type": "integer", "default": 10 },
          "include_metadata": { "type": "boolean", "default": false }
        }
      },
      "MemorySearchResult": {
        "type": "object",
        "properties": {
          "content": { "type": "string" },
          "score": { "type": "number" },
          "metadata": { "type": "object" }
        }
      },
...
    }
  }
}